* refactor: 폴더구조 재조정 1. components - shop/   - cart/   - product/ - admin/   - product/     - discount/   - coupon/ 2. utils 최상위로 변경
