+++ src/refactoring/components/admin/product/discount/AddNewDiscountForm.tsx
@@ -1,5 +1,5 @@
 import { useState } from 'react';
-import { Discount } from '../../../types.ts';
+import { Discount } from '../../../../../types.ts';
 
 const initialDiscount: Discount = {
   quantity: 0,
