+++ src/refactoring/utils/productUtils.ts
@@ -1,4 +1,4 @@
-import { Discount, Product } from "../../../types";
+import { Discount, Product } from "../../types";
 import { addDiscount, removeDiscount } from "./discountUtils";
 
 /**
